removed some warnings

+++ src/core/src/cnetwork_asio.cpp
@@ -211,7 +211,7 @@ bool CNetwork_Asio::Recv(uint16_t _size /*= 6*/) {
         asio::transfer_exactly(
             BytesToRead),  // We want at least 6 bytes of data
         [this](std::error_code errorCode, std::size_t length) {
-          packet_offset_ += length;
+          packet_offset_ += (uint16_t)length;
           last_update_time_ = (Core::Time::GetTickCount());
           if (!errorCode || errorCode.value() == 11) {
             if (OnReceived() == false) {
